mcp-server/requirements.txt
httpx
python-dotenv
mcp